pyproject.toml
[tool.poetry]
name = "ai_blitz"
version = "0.1.0"
description = ""
authors = ["Thomas Vetterli <thomas.vetterli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aicrowd-cli = "^0.1.14"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
